feat: modular openwebui generic function added
