+++ demo/ui/openwebui/src/lgos_openwebui/functions/generic/function.py
@@ -1,0 +1,10 @@
+"""
+title: Generic
+
+author: langgraph-openai-serve
+version: 0.25
+"""
+
+from .pipe import Pipe
+
+__all__ = ["Pipe"]
