+++ demo/ui/openwebui/tests/conftest.py
@@ -2,8 +2,8 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from lgos_openwebui.functions import generic
 from lgos_openwebui.functions.generic import Pipe
+from lgos_openwebui.functions.generic import pipe as generic_pipe
 
 from .openwebui_support import model
 
@@ -17,7 +17,7 @@ def anyio_backend() -> str:
 @pytest.fixture
 def configured_pipe(monkeypatch: pytest.MonkeyPatch) -> Pipe:
     monkeypatch.setattr(
-        generic,
+        generic_pipe,
         "_retrieve_model",
         AsyncMock(return_value=model()),
     )
